+++ Tests/TuistDependenciesTests/SwiftPackageManager/Utils/PackageInfoMapperTests.swift
@@ -1187,7 +1187,7 @@ final class PackageInfoMapperTests: TuistUnitTestCase {
             .test(
                 name: "Package",
                 targets: [
-                    .test("Target1", dependencies: [.xcframework(path: "/Path/To/Dependency1.framework")]),
+                    .test("Target1", dependencies: [.xcframework(path: "/artifacts/Package/Dependency1.xcframework")]),
                 ]
             )
         )
@@ -1227,7 +1227,7 @@ final class PackageInfoMapperTests: TuistUnitTestCase {
         )
     }
 
-    func testMap_whenBinaryTargetByNameDependency_mapsToXcFramework() throws {
+    func testMap_whenBinaryTargetURLByNameDependency_mapsToXcFramework() throws {
         let project = try subject.map(
             package: "Package",
             packageInfos: [
@@ -1238,9 +1238,11 @@ final class PackageInfoMapperTests: TuistUnitTestCase {
                     targets: [
                         .test(
                             name: "Target1",
-                            dependencies: [.byName(name: "Dependency1", condition: nil)]
+                            dependencies: [
+                                .byName(name: "Dependency1", condition: nil)
+                            ]
                         ),
-                        .test(name: "Dependency1", type: .binary),
+                        .test(name: "Dependency1", type: .binary), // url is default
                     ],
                     platforms: [],
                     cLanguageStandard: nil,
@@ -1249,7 +1251,7 @@ final class PackageInfoMapperTests: TuistUnitTestCase {
                 ),
             ],
             targetDependencyToFramework: [
-                "Dependency1": "/Path/To/Dependency1.framework",
+                "Dependency1": "/Path/To/Dependency1.framework"
             ]
         )
         XCTAssertEqual(
@@ -1257,7 +1259,45 @@ final class PackageInfoMapperTests: TuistUnitTestCase {
             .test(
                 name: "Package",
                 targets: [
-                    .test("Target1", dependencies: [.xcframework(path: "/Path/To/Dependency1.framework")]),
+                    .test("Target1", dependencies: [.xcframework(path: "/artifacts/Package/Dependency1.xcframework")])
+                ]
+            )
+        )
+    }
+    
+    func testMap_whenBinaryTargetPathByNameDependency_mapsToXcFramework() throws {
+        let project = try subject.map(
+            package: "Package",
+            packageInfos: [
+                "Package": .init(
+                    products: [
+                        .init(name: "Product1", type: .library(.automatic), targets: ["Target1"]),
+                    ],
+                    targets: [
+                        .test(
+                            name: "Target1",
+                            dependencies: [
+                                .byName(name: "Dependency1", condition: nil)
+                            ]
+                        ),
+                        .test(name: "Dependency1", type: .binary, path: "Dependency1/Dependency1.xcframework"),
+                    ],
+                    platforms: [],
+                    cLanguageStandard: nil,
+                    cxxLanguageStandard: nil,
+                    swiftLanguageVersions: nil
+                ),
+            ],
+            targetDependencyToFramework: [
+                "Dependency1": "Dependency1/Dependency1.framework",
+            ]
+        )
+        XCTAssertEqual(
+            project,
+            .test(
+                name: "Package",
+                targets: [
+                    .test("Target1", dependencies: [.xcframework(path: "Dependency1/Dependency1.xcframework")]),
                 ]
             )
         )
@@ -1535,9 +1575,10 @@ extension PackageInfoMapping {
                 result[product.name] = packageInfo.key
             }
         }
-
+        
+        let artifactsFolder = basePath.appending(component: "artifacts")
         let (targetToProducts, targetToResolvedDependencies) = try preprocess(
-            packageInfos: packageInfos, productToPackage: productToPackage, targetDependencyToFramework: targetDependencyToFramework
+            packageInfos: packageInfos, productToPackage: productToPackage, artifactsFolder: artifactsFolder
         )
 
         return try map(
@@ -1563,6 +1604,7 @@ extension PackageInfo.Target {
         name: String,
         type: PackageInfo.Target.TargetType = .regular,
         path: String? = nil,
+        url: String? = nil,
         sources: [String]? = nil,
         resources: [PackageInfo.Target.Resource] = [],
         dependencies: [PackageInfo.Target.Dependency] = [],
@@ -1572,7 +1614,7 @@ extension PackageInfo.Target {
         return .init(
             name: name,
             path: path,
-            url: nil,
+            url: url,
             sources: sources,
             resources: resources,
             exclude: [],